Refactor build scripts to import tools directly and streamline dependencies. Introduce new utility functions in tools.ts for better organization and maintainability.
