+++ scripts/prepare/bundle.ts
@@ -2,11 +2,13 @@ import { writeFile } from 'node:fs/promises';
 import { dirname, join, parse, posix, relative, sep } from 'node:path';
 
 import type { Metafile } from 'esbuild';
+import * as esbuild from 'esbuild';
 // eslint-disable-next-line depend/ban-dependencies
 import * as fs from 'fs-extra';
 // eslint-disable-next-line depend/ban-dependencies
 import { glob } from 'glob';
 import slash from 'slash';
+import { dedent } from 'ts-dedent';
 import type { Options } from 'tsup';
 import { build } from 'tsup';
 import type { PackageJson } from 'type-fest';
@@ -17,7 +19,7 @@ import {
   SUPPORTED_FEATURES,
 } from '../../code/core/src/shared/constants/environments-support';
 import { exec } from '../utils/exec';
-import { dedent, esbuild, nodeInternals } from './tools';
+import { nodeInternals } from '../utils/tools';
 
 /* TYPES */
